==> Projects/Permits/SALMON/analysis/08_rmd.Rmd <==
---
title: "08"
author: "Ben Buzzee"
date: "November 7, 2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
library(tidyverse)
```


```{r cars}
master_df_18 <- read.csv("master_df_18.csv")
master_df_18$year <- 2018


master_df_17 <- read.csv("master_df_17.csv")
master_df_17$year <- 2017

master_df_16 <- read.csv("master_df_16.csv")
master_df_16$year <- 2016

master_df <- rbind(master_df_18, master_df_17, master_df_16)
```



```{r}
# Did Not Fish
permits16 <- read.csv("../data/permit_records_16.csv") %>% filter(STATUS == "DID NOT FISH") %>% summarize(n()) %>% as.integer()
permits17 <- read.csv("../data/permit_records_17.csv") %>% filter(STATUS == "DID NOT FISH") %>% summarize(n()) %>% as.integer()
permits18 <- read.csv("../data/permit_records_18.csv") %>% filter(STATUS == "DID NOT FISH") %>% summarize(n()) %>% as.integer()




tab2row1 <- c(0, permits16 + permits17 + permits18, 0, 0)
tab3row1 <- c(0, permits16 + permits17 + permits18, 0, 0)
```




# Table 1: Permits across years

```{r}
master_df1 <- master_df %>% mutate(total_salmon = red + king + pink + coho + chum)
master_df1 <- master_df1 %>% group_by(permit, year, allowed, familysi) %>% summarize(total_salmon = sum(total_salmon),
                                                                           num_fisheries = length(unique(fishery)),
                                                                           days = sum(effort_days))

master_df1 <- master_df1 %>% mutate(perc_bag = total_salmon/allowed) %>% ungroup()

tab1 <- master_df1 %>% group_by(year) %>% summarize(total_permits = length((unique(permit))), total_salmon = sum(total_salmon), avg_perc_bag = mean(perc_bag))

knitr::kable(tab1)
```

# Table 2: Number of Fisheries Fished

```{r}

tab2 <- master_df1 %>% group_by(num_fisheries) %>% summarize(total_permits = length((unique(permit))), total_salmon = sum(total_salmon), avg_perc_bag = mean(perc_bag))

knitr::kable(rbind(tab2row1, tab2))

```

# Table 3: Number of Days Fished
```{r}
tab3 <- master_df1 %>% group_by(days) %>% summarize(total_permits = length((unique(permit))), total_salmon = sum(total_salmon), avg_perc_bag = mean(perc_bag))


knitr::kable(rbind(tab3row1, tab3))
```


# Table 4: Number of Household Members
```{r}
tab4 <- master_df1 %>% group_by(familysi) %>% summarize(total_permits = length((unique(permit))), total_salmon = sum(total_salmon), avg_perc_bag = mean(perc_bag))


knitr::kable(tab4)
```
